Make cards draggable and show drop zones
If a stack is empty always show the drop zone. Otherwise only show the
drop zone while a drag operation is ongoing, except for the source stack
from which the card was dragged.

Beware that the stacks and the cards must not be refreshed (e.g. by
changing a property monitored by Svelte), because this would interrupt
the drag operation and would require the user to drag the target element
a second time. We achieve this by limiting refreshes to the PlaceholderCard.