Calculate size of stacks based on child cards
Because the cards are positioned absolutely, the parent div sizes are
not automatically set to contain all children, but are 0, 0.

Exposing properties as CSS variable is borrowed from
https://svelte.dev/repl/14a1b548093642a9a2dfb3e615382732?version=3.48.0.
But as our interest is limited to width and height, we hard code the
variables in the setter functions, unlike the original code, which is
more generic.

KLUDGE(KNR): The card dimensions are currently hard coded in the stack
size functions. If possible they should be queries from a Card component
or read from a store. This will be fixed in a future commit.

The same approach will be applied to other container components like Box
in a future commit.

BEWARE(KNR): It's unclear whether the CSS variables will be updated,
should the size change dynamically. It all depends on whether the stack
div will be redrawn and whether the CSS setters are called in the
process.