refactor: Set child positions and size by properties
Set properties to the correct position and size either by the parent
component or by a reactive statement $:. Then convert those properties
into CSS variables and use the CSS variables in the CSS code.

Exposing properties as CSS variable is borrowed from
https://svelte.dev/repl/14a1b548093642a9a2dfb3e615382732?version=3.48.0.
But as our interest is limited to width and height, we hard code the
variables in the setter functions, unlike the original code, which is
more generic.

Factor out size calculations to display-helpers.js. Currently the card
dimensions are currently hard coded. If possible they should be queries
from a Card component or read from a store. This will be fixed in a future
commit.

Some helper functions are "overloaded" and either accept an array of
cards or the number of cards.

BEWARE(KNR): It's unclear whether the CSS variables will be updated,
should the size change dynamically. It all depends on whether the stack
div will be redrawn and whether the CSS setters are called in the
process.