feature: Visualize walls of open boxes using CSS 3D effects
This is the first step towards migrating from hand-rolled 3D effects by
offsetting X and Y coordinates to CSS 3D effects. That's why it does not
fit to the still manually positioned cards. Those will be migrated in a
future commit.

Also fixup the selection of boxes by managing box and stack data in the
app, which is the svelty way anyways. This way the app has the overview
of the boxes and can manage the selected box state, so that at most one
box is selected at once.