chore: Streamline units of position parameters
So far the function getUnit in display-helpers covered the most common
position parameters and appended 'px' when the position parameter was
applied through cssVariables. The uncommon position parameters required
explicit addition of the unit. By covering all position parameters in
getUnit that's no longer necessary. For this all position parameters are
passed as integers, which is ensured by assertions.