feature: Add closed box
Visualize a closed box as a box body with a slightly larger box lid. At
the front of the box display a label, even though currently we do not
support labelling boxes, yet.

To keep the positioning and size of the box body and the box lid simple,
add separate components for the body and the lid.

Currently it is not clear whether we need the backside of the box. For
now let's render it. Once we know how boxes are used we can potentially
remove them to reduce the rendering load.