chore: Generalize appending of units to variables
Before this commit a list of variable names had to be maintained in
display-helpers.js. To all variables matching one of those names the
unit 'px' was appended.

This approach has at least two drawbacks:
- it doesn't scale well, particularly when using more specific variable
names
- there is a chance that a particular variable should not be extended by
a unit even though the name of the variable is in the mentioned list

This commit generalizes the approach: all variables ending by '_px' will
be extended by unit 'px'.

New code will use this approach, and old code will be gradually migrated
or deleted.